Move out base types of predicates and events
