+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/event/types/OnNightStartEvent.kt
@@ -1,6 +1,7 @@
 package liltojustice.trueadaptivemusic.client.event.types
 
 import com.google.gson.JsonObject
+import liltojustice.trueadaptivemusic.client.event.MusicEvent
 
 class OnNightStartEvent(): MusicEvent() {
     companion object: MusicEventCompanion<OnNightStartEvent> {
