+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/MusicPredicate.kt
@@ -1,8 +1,8 @@
-package liltojustice.trueadaptivemusic.client.predicate.types
+package liltojustice.trueadaptivemusic.client.predicate
 
+import com.google.gson.JsonObject
 import liltojustice.trueadaptivemusic.client.MusicTrigger
 import liltojustice.trueadaptivemusic.client.ReflectionHelper
-import liltojustice.trueadaptivemusic.client.predicate.MusicPredicateException
 import net.minecraft.client.MinecraftClient
 import kotlin.reflect.KClass
 
@@ -12,6 +12,10 @@ abstract class MusicPredicate: MusicTrigger {
     companion object: MusicPredicateCompanion<MusicPredicate> {
         override fun getTypeName(): String {
             throw MusicPredicateException("Attempt to get type name from abstract predicate type.")
+        }
+
+        override fun fromJson(json: JsonObject): MusicPredicate {
+            return MusicTrigger.fromJsonProvideSubclasses(json, getTriggerImplementerSubclasses()) as MusicPredicate
         }
     }
