+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/DimensionPredicate.kt
@@ -3,6 +3,7 @@ package liltojustice.trueadaptivemusic.client.predicate.types
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import liltojustice.trueadaptivemusic.client.identifier.DimensionIdentifier
+import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import net.minecraft.client.MinecraftClient
 import net.minecraft.util.JsonHelper
