+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/HeightPredicate.kt
@@ -1,6 +1,7 @@
 package liltojustice.trueadaptivemusic.client.predicate.types
 
 import com.google.gson.JsonObject
+import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import net.minecraft.client.MinecraftClient
 import net.minecraft.util.JsonHelper
