+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/NightTimePredicate.kt
@@ -1,6 +1,7 @@
 package liltojustice.trueadaptivemusic.client.predicate.types
 
 import com.google.gson.JsonObject
+import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import net.minecraft.client.MinecraftClient
 
 class NightTimePredicate: MusicPredicate() {
