+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/StructurePredicate.kt
@@ -3,6 +3,7 @@ package liltojustice.trueadaptivemusic.client.predicate.types
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import liltojustice.trueadaptivemusic.client.identifier.StructureIdentifier
+import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import net.minecraft.client.MinecraftClient
 import net.minecraft.registry.RegistryKeys
 import net.minecraft.server.world.ServerWorld
