+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/WeatherPredicate.kt
@@ -2,6 +2,7 @@ package liltojustice.trueadaptivemusic.client.predicate.types
 
 import com.google.gson.JsonObject
 import liltojustice.trueadaptivemusic.client.enums.Weather
+import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import net.minecraft.client.MinecraftClient
 import net.minecraft.util.JsonHelper
